fix: address second round PR review comments for image augmentation
- CropNonEmptyMaskIfExists: check all channels for non-empty content
- DADA: replace greedy argmax with Gumbel-max trick for stochastic sampling
- HistogramEqualization: use long for totalPixels to prevent overflow
- HistogramMatching: add null check for reference histograms
- MotionBlur: fix odd kernel enforcement and use visited array for sum
- Perspective: return mask clone instead of reference
- Rain: fix halfWidth calculation and add params to GetParameters
- RandomBrightnessContrast: validate negative limit values
- Resize: clamp output after bicubic/Lanczos interpolation
- RgbToHls: set IsNormalized=true since HLS output is always [0,1]
- ThinPlateSpline: fix doc comment accuracy
- TokenMix: add maxMixRatio upper bound validation
- TransMix: update LastMixingLambda to reflect actual mixed ratio

+++ src/Augmentation/Image/HistogramEqualization.cs
@@ -62,7 +62,7 @@ public class HistogramEqualization<T> : ImageAugmenterBase<T>
     {
         // Build luminance histogram (average of all channels)
         var histogram = new int[numBins];
-        int totalPixels = h * w;
+        long totalPixels = (long)h * w;
 
         for (int y = 0; y < h; y++)
             for (int x = 0; x < w; x++)
@@ -110,7 +110,7 @@ public class HistogramEqualization<T> : ImageAugmenterBase<T>
         int h, int w, int channel, int numBins, double maxVal)
     {
         var histogram = new int[numBins];
-        int totalPixels = h * w;
+        long totalPixels = (long)h * w;
 
         // Build histogram
         for (int y = 0; y < h; y++)
